+++ platforms/unity/Assets/BarelyMusician/Scripts/SynthInstrument.cs
@@ -17,17 +17,17 @@ namespace Barely {
     }
 
     /// Envelope attack.
-    public float attack = 0.05f;
+    public double attack = 0.05;
 
     /// Envelope decay.
-    public float decay = 0.0f;
+    public double decay = 0.0;
 
     /// Envelope sustain.
     [Range(0.0f, 1.0f)]
-    public float sustain = 1.0f;
+    public double sustain = 1.0;
 
     /// Envelope release.
-    public float release = 0.25f;
+    public double release = 0.25;
 
     /// Oscillator type.
     public OscillatorType oscillatorType = OscillatorType.Sine;
@@ -41,8 +41,8 @@ namespace Barely {
       SetParameter(1, decay);
       SetParameter(2, sustain);
       SetParameter(3, release);
-      SetParameter(4, (float)oscillatorType);
-      SetParameter(5, (float)numVoices);
+      SetParameter(4, (double)oscillatorType);
+      SetParameter(5, (double)numVoices);
     }
   }
 }
